+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Unreleased
 
+* **Breaking** `.zr-rp` now cleans the values for some cases.
 * Fix `WINDOW.position().set` not moving the window.
 * Add `ZR_LICENSE` in `cargo zng res`.
 * **Breaking** Add `zng::env::About::license`.